06-workshops/03-AgentCore-identity/07-Outbound_Auth_3LO_ECS_Fargate/pyproject.toml
[project]
name = "amazon-ecs-with-agentcore-identity-and-3lo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = []

[tool.uv]
override-dependencies = [
    "urllib3>=2.6.3,<3"
]

[dependency-groups]
dev = [
    "aws-cdk-lib>=2.237.1",
    "cdk-nag>=2.30.0",
    "pytest>=8.0.0",
    "httpx>=0.27.0",
    "moto>=5.0.0",
    "checkov>=3.2.459",
    "ruff>=0.15.8",
]
oauth = [
    "bedrock-agentcore>=1.2.1",
    "cryptography>=44.0.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "PyJWT>=2.10.0",
    "uvicorn>=0.40.0",
]
runtime = [
    "bedrock-agentcore>=1.2.1",
    "fastapi>=0.128.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "strands-agents>=1.24.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
filterwarnings = [
    "ignore::pydantic.PydanticDeprecatedSince20:bedrock_agentcore.*",
]

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120

[tool.flake8]
max-line-length = 120
extend-ignore = ["E203", "W503"]

[tool.bandit]
exclude_dirs = ["tests", "cdk.out"]

[tool.pydocstyle]
convention = "google"
match-dir = '(?!tests|cdk\.out).*'
